+++ Adapter/PlentymarketsAdapter/Helper/ReferenceAmountCalculator.php
@@ -72,12 +72,6 @@ class ReferenceAmountCalculator implements ReferenceAmountCalculatorInterface
 
         $modifier = self::$convertionMatrix[$variationUnit]['conversion'];
 
-        $content = $variation['unit']['content'] * $modifier;
-
-        if ($content <= 0.25) {
-            return 0.1 / $modifier;
-        }
-
         return 1.0 / $modifier;
     }